+++ src/Recyclarr/Migration/IMigrationExecutor.cs
@@ -3,4 +3,5 @@ namespace Recyclarr.Migration;
 public interface IMigrationExecutor
 {
     void PerformAllMigrationSteps();
+    void CheckNeededMigrations();
 }