+++ src/Recyclarr/Migration/IMigrationStep.cs
@@ -5,6 +5,7 @@ public interface IMigrationStep
     int Order { get; }
     string Description { get; }
     IReadOnlyCollection<string> Remediation { get; }
+    bool Required { get; }
     bool CheckIfNeeded();
     void Execute();
 }